Move deprecation warnings tests to a module
