+++ testing/test_pluginmanager.py
@@ -374,15 +374,6 @@ def test_subset_hook_caller(pm):
     assert out == [10]
 
 
-def test_multicall_deprecated(pm):
-    class P1(object):
-        @hookimpl
-        def m(self, __multicall__, x):
-            pass
-
-    pytest.deprecated_call(pm.register, P1())
-
-
 def test_add_hookspecs_nohooks(pm):
     with pytest.raises(ValueError):
         pm.add_hookspecs(10)
@@ -409,23 +400,6 @@ def example_hook():
     with pytest.deprecated_call():
         assert pm.parse_hookimpl_opts(conftest, 'example_blah') is None
         assert pm.parse_hookimpl_opts(conftest, 'example_hook') == {}
-
-
-def test_callhistoric_proc_deprecated(pm):
-    """``proc`` kwarg to `PluginMananger.call_historic()` is now officially
-    deprecated.
-    """
-    class P1(object):
-        @hookspec(historic=True)
-        @hookimpl
-        def m(self, x):
-            pass
-
-    p1 = P1()
-    pm.add_hookspecs(p1)
-    pm.register(p1)
-    with pytest.deprecated_call():
-        pm.hook.m.call_historic(kwargs=dict(x=10), proc=lambda res: res)
 
 
 def test_load_setuptools_instantiation(monkeypatch, pm):
